flask_app/requirements.txt
Flask
flask-cors
flask-socketio
openai
python-dotenv
redis